requirements.txt
pygame==2.6.1
requests==2.31.0
python-dotenv==1.0.1
